Switch the read and dump subcommands to write and read, respectively; clean up the README
