+++ cmd/read.go
@@ -1,9 +1,8 @@
 package main
 
 import (
-	"flag"
 	"fmt"
-	"os"
+	"io/ioutil"
 )
 
 type ReadCommand struct {
@@ -11,50 +10,33 @@ type ReadCommand struct {
 }
 
 func (cmd *ReadCommand) Run() int {
-	fs := flag.NewFlagSet("list", flag.ContinueOnError)
-	clear := fs.Bool("clear", false, "overwrite contents")
-
-	fs.Usage = cmd.usage
-
-	if err := fs.Parse(cmd.args); err != nil {
-		cmd.usage()
-		return 1
-	}
-
-	cmd.args = fs.Args()
-
 	if len(cmd.args) < 2 {
 		cmd.usage()
 		return 1
 	}
 
-	reader := cmd.stdin
-
-	if cmd.args[1] != "-" {
-		file, err := os.Open(cmd.args[1])
-
-		if err != nil {
-			fmt.Fprintf(cmd.stderr, "read: %s\n", err)
-			return -1
-		}
-
-		defer file.Close()
-
-		reader = file
-	}
-
-	if *clear {
-		if err := cmd.Purge(cmd.args[0]); err != nil {
-			fmt.Fprintf(cmd.stderr, "read: %s\n", err)
-			return -1
-		}
-	}
-
-	err := cmd.Read(reader, cmd.args[0])
+	m, err := cmd.FetchAsJSON(cmd.args[0])
 
 	if err != nil {
-		fmt.Fprintf(cmd.stderr, "read: %s\n", err)
+		fmt.Fprintln(cmd.stderr, err)
 		return -1
+	}
+
+	// it's nice to have the file (or stdout)
+	// have a trailing newline
+
+	m = append(m, '\n')
+
+	if cmd.args[1] != "-" {
+		if err := ioutil.WriteFile(cmd.args[1], m, 0600); err != nil {
+			fmt.Fprintln(cmd.stderr, err)
+			return -1
+		}
+	} else {
+		if _, err := cmd.stdout.Write(m); err != nil {
+			fmt.Fprintln(cmd.stderr, err)
+			return -1
+		}
 	}
 
 	return 0
