+++ cmd/read_test.go
@@ -2,13 +2,15 @@ package main
 
 import (
 	"bytes"
+	"encoding/json"
 	"io/ioutil"
 	"os"
+	"path"
 	"reflect"
 	"testing"
 )
 
-func TestRead(t *testing.T) {
+func TestDump(t *testing.T) {
 	dname, err := ioutil.TempDir("", "scratch")
 
 	if err != nil {
@@ -17,57 +19,18 @@ func TestRead(t *testing.T) {
 
 	defer os.RemoveAll(dname)
 
-	file, err := ioutil.TempFile(dname, "test")
-
-	if err != nil {
-		t.Fatal("tempfile", err)
-	}
-
-	if _, err = file.WriteString(`{"x":"21", "y":"14"}`); err != nil {
-		t.Fatal("write-temp", err)
-	}
-
-	file.Close()
+	fname := path.Join(dname, "test.json")
 
 	stdout := new(bytes.Buffer)
 	stderr := new(bytes.Buffer)
 	app := NewTestApp(t, stdout, stderr)
+	expData := map[string]string{"a": "12", "b": "21"}
 
-	app.args = []string{"test", file.Name()}
-
-	cmd := ReadCommand{app}
-	o := cmd.Run()
-
-	if o != 0 {
-		t.Errorf("errors: %s", stderr.String())
-		t.Fatalf("invalid return: %d", o)
-	}
-
-	m, err := cmd.Fetch("test")
-
-	if err != nil {
-		t.Fatalf("can't fetch: %s", err)
-	}
-
-	exp := map[string]string{"x": "21", "y": "14"}
-
-	if !reflect.DeepEqual(m, exp) {
-		t.Errorf("invalid values: %#v", m)
-	}
-}
-
-func TestReadOverwrite(t *testing.T) {
-	stdout := new(bytes.Buffer)
-	stderr := new(bytes.Buffer)
-	app := NewTestApp(t, stdout, stderr)
-	data := map[string]string{"a": "xxx", "b": "yyy"}
-
-	if err := app.Add("test", data); err != nil {
+	if err := app.Add("test", expData); err != nil {
 		t.Fatal("setup", err)
 	}
 
-	app.stdin = bytes.NewBufferString(`{"x":"21", "y":"14"}`)
-	app.args = []string{"-clear", "test", "-"}
+	app.args = []string{"test", fname}
 
 	cmd := ReadCommand{app}
 	o := cmd.Run()
@@ -77,15 +40,52 @@ func TestReadOverwrite(t *testing.T) {
 		t.Fatalf("invalid return: %d", o)
 	}
 
-	m, err := cmd.Fetch("test")
+	raw, err := ioutil.ReadFile(fname)
 
 	if err != nil {
-		t.Fatalf("can't fetch: %s", err)
+		t.Fatal("read", err)
 	}
 
-	exp := map[string]string{"x": "21", "y": "14"}
+	t.Log(string(raw))
 
-	if !reflect.DeepEqual(m, exp) {
-		t.Errorf("invalid values: %#v", m)
+	var readData map[string]string
+
+	if err = json.Unmarshal(raw, &readData); err != nil {
+		t.Fatal("decode", err)
+	}
+
+	if !reflect.DeepEqual(readData, expData) {
+		t.Errorf("invalid data: %#v", readData)
+	}
+}
+
+func TestDumpStdout(t *testing.T) {
+	stdout := new(bytes.Buffer)
+	stderr := new(bytes.Buffer)
+	app := NewTestApp(t, stdout, stderr)
+	expData := map[string]string{"a": "12", "b": "21"}
+
+	if err := app.Add("test", expData); err != nil {
+		t.Fatal("setup", err)
+	}
+
+	app.args = []string{"test", "-"}
+
+	cmd := ReadCommand{app}
+	o := cmd.Run()
+
+	if o != 0 {
+		t.Errorf("errors: %s", stderr.String())
+		t.Fatalf("invalid return: %d", o)
+	}
+
+	var readData map[string]string
+
+	if err := json.NewDecoder(stdout).Decode(&readData); err != nil {
+		t.Fatal("decode", err)
+	}
+
+	if !reflect.DeepEqual(readData, expData) {
+		t.Errorf("invalid data: %#v", readData)
 	}
 }
